dual approval refactoring
